use float for division in scale computation

+++ src/DigitalBaro/WeatherLcdGraph.cpp
@@ -17,7 +17,7 @@ void WeatherLcdGraph::setLimits(int16_t min, int16_t max)
 {
   m_minY = min;
   m_maxY = max;
-  m_scale = (max-min)/GRAPH_Y_PIXELS;
+  m_scale = (max-min)/(float)GRAPH_Y_PIXELS;
 }
 
 int16_t WeatherLcdGraph::minY() {
